+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Chokidar Changelog
 
+### 1.7.2
+* Better handling of renamed and removed directories
+* Better handling of paths when cwd is submitted
+
 ### 1.7.1
 * Continuous integration
 * Bumped anymatch version to patch minor vulnerability